backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
httpx
python-dotenv
